Sesion 11

+++ 7_Token refreshing and Flask-JWT-Extended/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required,get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 class Item(Resource):
@@ -15,7 +15,7 @@ class Item(Resource):
             return item.json()
         return {'message':'item not found'},404
     
-    
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message":f"item: {name} already exist"}
@@ -32,7 +32,11 @@ class Item(Resource):
 
         return item.json(), 201 
     
+    @jwt_required
     def delete(self,name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return{"message":"Admin privilege required"}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -52,5 +56,10 @@ class Item(Resource):
         return item.json()
     
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {"items":items},200
+        return {'items': [item['name'] for item in items],"message":"more data avalible if you login"}
